+++ appcontainer/AppContainerModel.js
@@ -4,10 +4,12 @@
  *
  * Copyright © 2021 Extremely Heavy Industries Inc.
  */
-import {HoistModel, managed, RootRefreshContextModel} from '@xh/hoist/core';
-import {action, observable, makeObservable} from '@xh/hoist/mobx';
+import {XH, HoistModel, managed, RootRefreshContextModel} from '@xh/hoist/core';
 import {PendingTaskModel} from '@xh/hoist/utils/async';
+import {Icon} from '@xh/hoist/icon';
+
 import {AboutDialogModel} from './AboutDialogModel';
+import {BannerSourceModel} from './BannerSourceModel';
 import {ExceptionDialogModel} from './ExceptionDialogModel';
 import {FeedbackDialogModel} from './FeedbackDialogModel';
 import {ImpersonationBarModel} from './ImpersonationBarModel';
@@ -31,6 +33,7 @@ export class AppContainerModel extends HoistModel {
     @managed impersonationBarModel = new ImpersonationBarModel();
     @managed messageSourceModel = new MessageSourceModel();
     @managed toastSourceModel = new ToastSourceModel();
+    @managed bannerSourceModel = new BannerSourceModel();
     @managed themeModel = new ThemeModel();
     @managed refreshContextModel = new RootRefreshContextModel();
 
@@ -40,11 +43,6 @@ export class AppContainerModel extends HoistModel {
      */
     @managed
     appLoadModel = new PendingTaskModel({mode: 'all'});
-
-    constructor() {
-        super();
-        makeObservable(this);
-    }
 
     init() {
         const models = [
@@ -64,25 +62,35 @@ export class AppContainerModel extends HoistModel {
         });
     }
 
-    /** Updated App version available, as reported by server. */
-    @observable updateVersion = null;
-
     /**
-     * Show the update toolbar prompt. Called by EnvironmentService when the server reports that a
+     * Show the update Banner. Called by EnvironmentService when the server reports that a
      * new (or at least different) version is available and the user should be prompted.
      *
      * @param {string} version - updated version from server.
      * @param {string} [build] - updated build from server - included for snapshot version prompts.
      */
-    @action
-    showUpdateBar(version, build) {
-        let updateVersion = version;
-
+    showUpdateBanner(version, build) {
         // Display build tag for snaps only - not of much interest across actual version updates.
-        if (updateVersion.includes('SNAPSHOT') && build && build != 'UNKNOWN') {
-            updateVersion += ` (b${build})`;
+        if (version.includes('SNAPSHOT') && build && build !== 'UNKNOWN') {
+            version += ` (b${build})`;
         }
 
-        this.updateVersion = updateVersion;
+        // Show Banner
+        const mobile = XH.isMobileApp,
+            message = mobile ? 'Update available!' : `A new version of ${XH.clientAppName} is available!`,
+            buttonText = mobile ? version : `Update to ${version}`;
+
+        XH.showBanner({
+            category: 'app-update',
+            message,
+            icon: Icon.rocket({size: 'lg'}),
+            intent: 'warning',
+            enableClose: false,
+            actionButtonProps: {
+                icon: Icon.refresh(),
+                text: buttonText,
+                onClick: () => XH.reloadApp()
+            }
+        });
     }
 }