+++ desktop/appcontainer/AppContainer.js
@@ -4,7 +4,6 @@
  *
  * Copyright © 2021 Extremely Heavy Industries Inc.
  */
-import {hotkeysProvider} from '@xh/hoist/kit/blueprint';
 import {AppContainerModel} from '@xh/hoist/appcontainer/AppContainerModel';
 import {fragment, frame, vframe, viewport} from '@xh/hoist/cmp/layout';
 import {AppState, elem, hoistCmp, refreshContextView, uses, XH} from '@xh/hoist/core';
@@ -14,13 +13,16 @@ import {dockContainerImpl} from '@xh/hoist/desktop/cmp/dock/impl/DockContainer';
 import {colChooserDialog as colChooser} from '@xh/hoist/desktop/cmp/grid/impl/ColChooserDialog';
 import {ColChooserModel} from '@xh/hoist/desktop/cmp/grid/impl/ColChooserModel';
 import {mask} from '@xh/hoist/desktop/cmp/mask';
+import {pinPadImpl} from '@xh/hoist/desktop/cmp/pinpad/impl/PinPad';
 import {storeFilterFieldImpl} from '@xh/hoist/desktop/cmp/store/impl/StoreFilterField';
 import {tabContainerImpl} from '@xh/hoist/desktop/cmp/tab/impl/TabContainer';
-import {pinPadImpl} from '@xh/hoist/desktop/cmp/pinpad/impl/PinPad';
-import {useHotkeys, useContextMenu} from '@xh/hoist/desktop/hooks';
+import {useContextMenu, useHotkeys} from '@xh/hoist/desktop/hooks';
 import {installDesktopImpls} from '@xh/hoist/dynamics/desktop';
-import {useOnMount, elementFromContent} from '@xh/hoist/utils/react';
+import {hotkeysProvider} from '@xh/hoist/kit/blueprint';
+import {elementFromContent, useOnMount} from '@xh/hoist/utils/react';
+import {isEmpty} from 'lodash';
 import {aboutDialog} from './AboutDialog';
+import {banner} from './Banner';
 import {exceptionDialog} from './ExceptionDialog';
 import {feedbackDialog} from './FeedbackDialog';
 import {idlePanel} from './IdlePanel';
@@ -30,7 +32,6 @@ import {loginPanel} from './LoginPanel';
 import {messageSource} from './MessageSource';
 import {optionsDialog} from './OptionsDialog';
 import {toastSource} from './ToastSource';
-import {updateBar} from './UpdateBar';
 import {versionBar} from './VersionBar';
 
 installDesktopImpls({
@@ -48,7 +49,7 @@ installDesktopImpls({
  * This class provide core Hoist Application layout and infrastructure to an application's
  * root Component. Provides a standard viewport that includes standard UI elements such as an
  * impersonation bar header, version bar footer, an app-wide load mask, a base context menu,
- * popup message support, and exception rendering.
+ * popup+banner message support, and exception rendering.
  *
  * This component will kick off the Hoist application lifecycle when mounted.
  */
@@ -106,7 +107,7 @@ const appContainerView = hoistCmp.factory({
         let ret = viewport(
             vframe(
                 impersonationBar(),
-                updateBar(),
+                bannerList(),
                 refreshContextView({
                     model: model.refreshContextModel,
                     item: frame(elem(appSpec.componentClass, {model: appModel}))
@@ -126,6 +127,16 @@ const appContainerView = hoistCmp.factory({
         ret = useHotkeys(ret, globalHotKeys(model));
 
         return ret;
+    }
+});
+
+const bannerList = hoistCmp.factory({
+    render({model}) {
+        const {bannerModels} = model.bannerSourceModel;
+        if (isEmpty(bannerModels)) return null;
+        return fragment({
+            items: bannerModels.map(model => banner({model, key: model.xhId}))
+        });
     }
 });